Add explicit type cast
This was thought to be needed as part of an investigation and resolution for setting the permissions, a.k.a. modes, for a file. There was confusion whether permissions were stored as signed or unsigned long values within the libssh2 library. The explicit cast was added to ensure the type matched the libssh2 function signature.

+++ src/lv-libssh2-sftp.c
@@ -418,7 +418,7 @@ lv_libssh2_sftp_file_rename(
         (unsigned int)strlen(source_path),
         destination_path,
         (unsigned int)strlen(destination_path),
-        options
+        (long)options
     );
     if (result != 0) {
         return lv_libssh2_sftp_status_from_result(handle->inner, result);
